+++ serverImg.py
@@ -23,16 +23,25 @@ def default():
 
 @app.route('/analyze',methods=['POST'])
 def detectImage():
-    files = list(request.files)
-    results=[]
-    for file in files:
-        tmp=str(math.floor(random.random()*1000000))+ request.files[file].filename
-        request.files[file].save('images/'+tmp)
-        result=yolo.detectimage('images/'+tmp)
-        os.remove('images/'+tmp)
-        results.append({tmp:result[1]})
+    if request.form.get('path')=='true':
+        results=[]
+        path = request.form.get('filepath')
+        result=yolo.detectimage(path,True)
+        results.append({path: result[1]})
         print(result[1])
-        collection.insert_one({'filename':tmp,'data':result[1]})
-    return {'response':results}
+        collection.insert_one({'filename': path, 'data': result[1]})
+        return {'response': results}
+    else:
+        files = list(request.files)
+        results=[]
+        for file in files:
+            tmp=str(math.floor(random.random()*1000000))+ request.files[file].filename
+            request.files[file].save('images/'+tmp)
+            result=yolo.detectimage('images/'+tmp,False)
+            os.remove('images/'+tmp)
+            results.append({tmp:result[1]})
+            print(result[1])
+            collection.insert_one({'filename':tmp,'data':result[1]})
+        return {'response':results}
 
 app.run(host=host,port=imgport)
